+++ src/utils/postprocessing/EffectComposer.js
